refactor: move tests to trigger file
